pyproject.toml
[tool.poetry]
name = "todoer"
version = "0.1.0"
description = "Removes todo that match a certain regex."
authors = ["Max Wittig <max.wittig95@gmail.com>"]
readme = "README.md"
repository = "https://github.com/max-wittig/todoer"
keywords = ["gitlab", "todo"]
classifiers=[
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Topic :: Software Development",
    "Topic :: Utilities"
]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
loguru = "^0.4.1"
python-gitlab = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[tool.poetry.scripts]
todoer = "todoer.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
